Commented out unused codes for stripe, deleted empty and unused files, and moved cart files to their own components.
