+++ client/src/components/AddProduct/index.js
@@ -1,5 +1,5 @@
-import React from "react";
+// import React from "react";
 
-export default function AddProduct() {
-  return <>AddProduct</>
-}
+// export default function AddProduct() {
+//   return <>AddProduct</>
+// }
