+++ client/src/components/Cart/index.js
@@ -1,59 +1,60 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
-import CartItem from './CartItem';
+import { makeStyles} from '@material-ui/core/styles';
+import {Card, CardActionArea, CardActions, CardContent, CardMedia , IconButton,
+Typography, Checkbox, FormControlLabel  } from '@material-ui/core';
+import RemoveShoppingCartIcon from '@material-ui/icons/RemoveShoppingCart';
+import CheckBoxIcon from '@material-ui/icons/CheckBox';
+import CheckBoxOutlineBlankIcon from '@material-ui/icons/CheckBoxOutlineBlank';
 
-export default class Cart extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = { products: [], total: 0 }
-    }
+const useStyles = makeStyles({
+  root: {
+    width: "auto",
+    margin: "10px"
+  },
+  media: {
+    height: "150px",
+  },
+  content: {
+      padding: "10px"
+  },
+});
 
-    componentDidMount() {
-        let cart = localStorage.getItem('cart');
-        if (!cart) return;
-        getCartProducts(cart).then((products) => {
-            let total = 0;
-            for (var i = 0; i < products.length; i++) {
-                total += products[i].price * products[i].qty;
-            }
-            this.setState({ products, total });
-            });
-    }
+export default function CartCard() {
+  const classes = useStyles();
 
-    removeFromCart = (products) => {
-        let products = this.state.products.filter((item) => item.id !== product.id);
-        let cart = JSON.parse(localStorage.getItem('cart'));
-        delete cart[product.id.toString()];
-        localStorage.setItem('cart', JSON.stringify(cart));
-        let total = this.state.total - (product.qty * product.price)
-        this.setState({ products, total });
-    }
+  return (
+    <Card className={classes.root}>
+      <CardActionArea>
+        <CardMedia
+          className={classes.media}
+          image="https://images.pexels.com/photos/712367/pexels-photo-712367.jpeg?auto=compress&cs=tinysrgb&dpr=3&h=750&w=1260"
+          title="product"
+        />
+        <CardContent className={classes.content}>
+          <Typography gutterBottom variant="h5" component="h2">
+            Jelly fish named Jelly
+          </Typography>
+          <Typography variant="body2" color="textSecondary" component="p">
+            $ 4.91
+          </Typography>
+        </CardContent>
+      </CardActionArea>
+      <CardActions style={{backgroundColor:"lightgrey", maxHeight:"15px"}}>
+         <FormControlLabel
+        control={
+          <Checkbox
+            icon={<CheckBoxOutlineBlankIcon fontSize="small" />}
+            checkedIcon={<CheckBoxIcon fontSize="small" />}
+            name="checkedI"
+          />
+        }
+        label="Gift?"
+      />
 
-    clearCart = () => {
-        localStorage.removeItem('cart');
-        this.setState({ products: [] });
-    }
-
-    render() {
-        const { products, total } = this.state;
-        return (
-            <div className="container">
-                <h3 className="card-title">Cart</h3><hr />
-                {
-                    products.map((product, index) =>
-                        <CartItem product={product} remove={this.removeFromCart} key={index} />)
-                }<hr />
-                { products.length ?
-                    <div><h4>
-                        <small>Total Amount: </small>
-                        <span className="float-right text-primary">${total}</span>
-                    </h4><hr /></div> : ''}
-                { !products.length ? <h3 className="text-warning">No item in the cart</h3> : ''}
-                <Link to="/checkout">
-                    <button className="btn btn-success float-right">Checkout</button></Link>
-                <button className="btn btn-danger float-right" onClick={this.clearCart}
-                    style={{ marginRight: "10px" }}>Clear Cart</button><br /><br /><br />
-            </div>
-        );
-    }
+        <IconButton aria-label="delete"className={classes.delete}>
+          <RemoveShoppingCartIcon />
+        </IconButton>
+      </CardActions>
+    </Card>
+  );
 }
